warning avoided

+++ ql/MonteCarlo/montecarlomodel.hpp
@@ -93,7 +93,7 @@ namespace QuantLib {
         template<class S, class PG, class PP>
         inline void MonteCarloModel<S, PG, PP>::
                     addSamples(unsigned int samples) {
-            for(long j = 1; j <= samples; j++) {
+            for(unsigned int j = 1; j <= samples; j++) {
                 typename PG::sample_type path = pathGenerator_->next();
                 typename PP::result_type price = (*pathPricer_)(path);
                 double weight = pathGenerator_->weight();
